Django/api-access-control-system/requirements.txt
django==3.2.11
djangorestframework==3.13.1